# contracts/openapi/openapi.yaml
...
info:
  title: SEO Blog CMS API
  version: 0.1.0
openapi: 3.1.0
paths:
...
  /content/v1:
    delete:
      operationId: delete_content_v1
      responses:
        "200":
          description: Successful response
        "400":
          description: Invalid request
        "401":
          description: Authentication required
        "403":
          description: Insufficient permission
        "404":
          description: Resource not found
        "500":
          description: Internal server error
      summary: DELETE /content/v1
      tags:
        - Content API
    get:
      operationId: get_content_v1
      responses:
        "200":
          description: Successful response
        "400":
          description: Invalid request
        "401":
          description: Authentication required
        "403":
          description: Insufficient permission
        "404":
          description: Resource not found
        "500":
          description: Internal server error
      summary: GET /content/v1
      tags:
        - Content API
    head:
      operationId: head_content_v1
      responses:
        "200":
          description: Successful response
        "400":
          description: Invalid request
        "401":
          description: Authentication required
        "403":
          description: Insufficient permission
        "404":
          description: Resource not found
        "500":
          description: Internal server error
      summary: HEAD /content/v1
      tags:
        - Content API
    options:
      operationId: options_content_v1
      responses:
        "200":
          description: Successful response
        "400":
          description: Invalid request
        "401":
          description: Authentication required
        "403":
          description: Insufficient permission
        "404":
          description: Resource not found
        "500":
          description: Internal server error
      summary: OPTIONS /content/v1
      tags:
        - Content API
    patch:
      operationId: patch_content_v1
      responses:
        "200":
          description: Successful response
        "400":
          description: Invalid request
        "401":
          description: Authentication required
        "403":
          description: Insufficient permission
        "404":
          description: Resource not found
        "500":
          description: Internal server error
      summary: PATCH /content/v1
      tags:
        - Content API
    post:
      operationId: post_content_v1
      responses:
        "200":
          description: Successful response
        "400":
          description: Invalid request
        "401":
          description: Authentication required
        "403":
          description: Insufficient permission
        "404":
          description: Resource not found
        "500":
          description: Internal server error
      summary: POST /content/v1
      tags:
        - Content API
    put:
      operationId: put_content_v1
      responses:
        "200":
          description: Successful response
        "400":
          description: Invalid request
        "401":
          description: Authentication required
        "403":
          description: Insufficient permission
        "404":
          description: Resource not found
        "500":
          description: Internal server error
      summary: PUT /content/v1
      tags:
        - Content API
  /content/v1/authors:
    get:
      operationId: get_content_v1_authors
      responses:
        "200":
          description: Successful response
        "400":
          description: Invalid request
        "401":
          description: Authentication required
        "403":
          description: Insufficient permission
        "404":
          description: Resource not found
        "500":
          description: Internal server error
      summary: GET /content/v1/authors
      tags:
        - Content API
    head:
      operationId: head_content_v1_authors
      responses:
        "200":
          description: Successful response
        "400":
          description: Invalid request
        "401":
          description: Authentication required
        "403":
          description: Insufficient permission
        "404":
          description: Resource not found
        "500":
          description: Internal server error
      summary: HEAD /content/v1/authors
      tags:
        - Content API
  /content/v1/authors/{slug}:
    get:
      operationId: get_content_v1_authors_slug
      responses:
        "200":
          description: Successful response
        "400":
          description: Invalid request
        "401":
          description: Authentication required
        "403":
          description: Insufficient permission
        "404":
          description: Resource not found
        "500":
          description: Internal server error
      summary: GET /content/v1/authors/{slug}
      tags:
        - Content API
    head:
      operationId: head_content_v1_authors_slug
      responses:
        "200":
          description: Successful response
        "400":
          description: Invalid request
        "401":
          description: Authentication required
        "403":
          description: Insufficient permission
        "404":
          description: Resource not found
        "500":
          description: Internal server error
      summary: HEAD /content/v1/authors/{slug}
      tags:
        - Content API
  /content/v1/categories:
    get:
      operationId: get_content_v1_categories
      responses:
        "200":
          description: Successful response
        "400":
          description: Invalid request
        "401":
          description: Authentication required
        "403":
          description: Insufficient permission
        "404":
          description: Resource not found
        "500":
          description: Internal server error
      summary: GET /content/v1/categories
      tags:
        - Content API
    head:
      operationId: head_content_v1_categories
      responses:
        "200":
          description: Successful response
        "400":
          description: Invalid request
        "401":
          description: Authentication required
        "403":
          description: Insufficient permission
        "404":
          description: Resource not found
        "500":
          description: Internal server error
      summary: HEAD /content/v1/categories
      tags:
        - Content API
  /content/v1/categories/{slug}:
    get:
      operationId: get_content_v1_categories_slug
      responses:
        "200":
          description: Successful response
        "400":
          description: Invalid request
        "401":
          description: Authentication required
        "403":
          description: Insufficient permission
        "404":
          description: Resource not found
        "500":
          description: Internal server error
      summary: GET /content/v1/categories/{slug}
      tags:
        - Content API
    head:
      operationId: head_content_v1_categories_slug
      responses:
        "200":
          description: Successful response
        "400":
          description: Invalid request
        "401":
          description: Authentication required
        "403":
          description: Insufficient permission
        "404":
          description: Resource not found
        "500":
          description: Internal server error
      summary: HEAD /content/v1/categories/{slug}
      tags:
        - Content API
  /content/v1/changes:
    get:
      operationId: get_content_v1_changes
      responses:
        "200":
          description: Successful response
        "400":
          description: Invalid request
        "401":
          description: Authentication required
        "403":
          description: Insufficient permission
        "404":
          description: Resource not found
        "500":
          description: Internal server error
      summary: GET /content/v1/changes
      tags:
        - Content API
    head:
      operationId: head_content_v1_changes
      responses:
        "200":
          description: Successful response
        "400":
          description: Invalid request
        "401":
          description: Authentication required
        "403":
          description: Insufficient permission
        "404":
          description: Resource not found
        "500":
          description: Internal server error
      summary: HEAD /content/v1/changes
      tags:
        - Content API
  /content/v1/discovery-manifest:
    get:
      operationId: get_content_v1_discovery_manifest
      responses:
        "200":
          description: Successful response
        "400":
          description: Invalid request
        "401":
          description: Authentication required
        "403":
          description: Insufficient permission
        "404":
          description: Resource not found
        "500":
          description: Internal server error
      summary: GET /content/v1/discovery-manifest
      tags:
        - Content API
    head:
      operationId: head_content_v1_discovery_manifest
      responses:
        "200":
          description: Successful response
        "400":
          description: Invalid request
        "401":
          description: Authentication required
        "403":
          description: Insufficient permission
        "404":
          description: Resource not found
        "500":
          description: Internal server error
      summary: HEAD /content/v1/discovery-manifest
      tags:
        - Content API
  /content/v1/feed-data:
    get:
      operationId: get_content_v1_feed_data
      responses:
        "200":
          description: Successful response
        "400":
          description: Invalid request
        "401":
          description: Authentication required
        "403":
          description: Insufficient permission
        "404":
          description: Resource not found
        "500":
          description: Internal server error
      summary: GET /content/v1/feed-data
      tags:
        - Content API
    head:
      operationId: head_content_v1_feed_data
      responses:
        "200":
          description: Successful response
        "400":
          description: Invalid request
        "401":
          description: Authentication required
        "403":
          description: Insufficient permission
        "404":
          description: Resource not found
        "500":
          description: Internal server error
      summary: HEAD /content/v1/feed-data
      tags:
        - Content API
  /content/v1/posts:
    get:
      operationId: get_content_v1_posts
      responses:
        "200":
          description: Successful response
        "400":
          description: Invalid request
        "401":
          description: Authentication required
        "403":
          description: Insufficient permission
        "404":
          description: Resource not found
        "500":
          description: Internal server error
      summary: GET /content/v1/posts
      tags:
        - Content API
    head:
      operationId: head_content_v1_posts
      responses:
        "200":
          description: Successful response
        "400":
          description: Invalid request
        "401":
          description: Authentication required
        "403":
          description: Insufficient permission
        "404":
          description: Resource not found
        "500":
          description: Internal server error
      summary: HEAD /content/v1/posts
      tags:
        - Content API
  /content/v1/posts/by-id/{contentID}:
    get:
      operationId: get_content_v1_posts_by_id_contentid
      responses:
        "200":
          description: Successful response
        "400":
          description: Invalid request
        "401":
          description: Authentication required
        "403":
          description: Insufficient permission
        "404":
          description: Resource not found
        "500":
          description: Internal server error
      summary: GET /content/v1/posts/by-id/{contentID}
      tags:
        - Content API
    head:
      operationId: head_content_v1_posts_by_id_contentid
      responses:
        "200":
          description: Successful response
        "400":
          description: Invalid request
        "401":
          description: Authentication required
        "403":
          description: Insufficient permission
        "404":
          description: Resource not found
        "500":
          description: Internal server error
      summary: HEAD /content/v1/posts/by-id/{contentID}
      tags:
        - Content API
  /content/v1/posts/{slug}:
    get:
      operationId: get_content_v1_posts_slug
      responses:
        "200":
          description: Successful response
        "400":
          description: Invalid request
        "401":
          description: Authentication required
        "403":
          description: Insufficient permission
        "404":
          description: Resource not found
        "500":
          description: Internal server error
      summary: GET /content/v1/posts/{slug}
      tags:
        - Content API
    head:
      operationId: head_content_v1_posts_slug
      responses:
        "200":
          description: Successful response
        "400":
          description: Invalid request
        "401":
          description: Authentication required
        "403":
          description: Insufficient permission
        "404":
          description: Resource not found
        "500":
          description: Internal server error
      summary: HEAD /content/v1/posts/{slug}
      tags:
        - Content API
  /content/v1/posts/{slug}/related:
    get:
      operationId: get_content_v1_posts_slug_related
      responses:
        "200":
          description: Successful response
        "400":
          description: Invalid request
        "401":
          description: Authentication required
        "403":
          description: Insufficient permission
        "404":
          description: Resource not found
        "500":
          description: Internal server error
      summary: GET /content/v1/posts/{slug}/related
      tags:
        - Content API
    head:
      operationId: head_content_v1_posts_slug_related
      responses:
        "200":
          description: Successful response
        "400":
          description: Invalid request
        "401":
          description: Authentication required
        "403":
          description: Insufficient permission
        "404":
          description: Resource not found
        "500":
          description: Internal server error
      summary: HEAD /content/v1/posts/{slug}/related
      tags:
        - Content API
  /content/v1/redirects:
    get:
      operationId: get_content_v1_redirects
      responses:
        "200":
          description: Successful response
        "400":
          description: Invalid request
        "401":
          description: Authentication required
        "403":
          description: Insufficient permission
        "404":
          description: Resource not found
        "500":
          description: Internal server error
      summary: GET /content/v1/redirects
      tags:
        - Content API
    head:
      operationId: head_content_v1_redirects
      responses:
        "200":
          description: Successful response
        "400":
          description: Invalid request
        "401":
          description: Authentication required
        "403":
          description: Insufficient permission
        "404":
          description: Resource not found
        "500":
          description: Internal server error
      summary: HEAD /content/v1/redirects
      tags:
        - Content API
  /content/v1/series:
    get:
      operationId: get_content_v1_series
      responses:
        "200":
          description: Successful response
        "400":
          description: Invalid request
        "401":
          description: Authentication required
        "403":
          description: Insufficient permission
        "404":
          description: Resource not found
        "500":
          description: Internal server error
      summary: GET /content/v1/series
      tags:
        - Content API
    head:
      operationId: head_content_v1_series
      responses:
        "200":
          description: Successful response
        "400":
          description: Invalid request
        "401":
          description: Authentication required
        "403":
          description: Insufficient permission
        "404":
          description: Resource not found
        "500":
          description: Internal server error
      summary: HEAD /content/v1/series
      tags:
        - Content API
  /content/v1/series/{slug}:
    get:
      operationId: get_content_v1_series_slug
      responses:
        "200":
          description: Successful response
        "400":
          description: Invalid request
        "401":
          description: Authentication required
        "403":
          description: Insufficient permission
        "404":
          description: Resource not found
        "500":
          description: Internal server error
      summary: GET /content/v1/series/{slug}
      tags:
        - Content API
    head:
      operationId: head_content_v1_series_slug
      responses:
        "200":
          description: Successful response
        "400":
          description: Invalid request
        "401":
          description: Authentication required
        "403":
          description: Insufficient permission
        "404":
          description: Resource not found
        "500":
          description: Internal server error
      summary: HEAD /content/v1/series/{slug}
      tags:
        - Content API
...